backend/requirements.txt
asgiref
attrs
certifi
cffi
charset-normalizer
coreapi
coreschema
cryptography
defusedxml
Django
django-extra-fields
django-filter
djangorestframework
flake8-plugin-utils
flake8-polyfill
gunicorn
idna
inflection
iniconfig
isort
itypes
Jinja2
MarkupSafe
mccabe
oauthlib
packaging
Pillow
pluggy
psycopg2-binary
py
pycodestyle
pycparser
pyflakes
PyJWT
pyparsing
pytest
pytest-django
python-environ
python3-openid
pytz
reportlab
requests
requests-oauthlib
ruamel.yaml
ruamel.yaml.clib
six
sqlparse
testfixtures
toml
uritemplate
urllib3
